Moving relevant tests into the mixin test file Fixing test structure for new mixin format
