+++ tests/helcim/gateway/test_gateway_base_card_transaction.py
@@ -1,17 +1,7 @@
 """Tests for the gateway module."""
 # pylint: disable=missing-docstring, protected-access
-from datetime import date
-from unittest.mock import patch
+from helcim import gateway
 
-from helcim import gateway, exceptions as helcim_exceptions
-
-
-class MockDjangoModel():
-    def __init__(self, **kwargs):
-        self.data = kwargs
-
-def mock_get_or_create_created(**kwargs):
-    return (MockDjangoModel(**kwargs), True)
 
 API_DETAILS = {
     'url': 'https://www.test.com',
@@ -19,66 +9,6 @@ API_DETAILS = {
     'token': 'abcdefg',
     'terminal_id': '98765432',
 }
-
-@patch.dict('helcim.gateway.SETTINGS', {'enable_token_vault': True})
-def test_determine_save_token_status_enabled_user_yes():
-    details = {
-        'token': 'abcdefghijklmnopqrstuvw',
-        'customer_code': 'CST1000',
-        'token_f4l4': '11119999',
-    }
-
-    transaction = gateway.BaseCardTransaction(
-        api_details=API_DETAILS, **details
-    )
-    status = transaction._determine_save_token_status(True)
-
-    assert status is True
-
-@patch.dict('helcim.gateway.SETTINGS', {'enable_token_vault': True})
-def test_determine_save_token_status_enabled_user_no():
-    details = {
-        'token': 'abcdefghijklmnopqrstuvw',
-        'customer_code': 'CST1000',
-        'token_f4l4': '11119999',
-    }
-
-    transaction = gateway.BaseCardTransaction(
-        api_details=API_DETAILS, **details
-    )
-    status = transaction._determine_save_token_status(False)
-
-    assert status is False
-
-@patch.dict('helcim.gateway.SETTINGS', {'enable_token_vault': False})
-def test_determine_save_token_status_disabled_user_yes():
-    details = {
-        'token': 'abcdefghijklmnopqrstuvw',
-        'customer_code': 'CST1000',
-        'token_f4l4': '11119999',
-    }
-
-    transaction = gateway.BaseCardTransaction(
-        api_details=API_DETAILS, **details
-    )
-    status = transaction._determine_save_token_status(True)
-
-    assert status is False
-
-@patch.dict('helcim.gateway.SETTINGS', {'enable_token_vault': False})
-def test_determine_save_token_status_disabled_user_no():
-    details = {
-        'token': 'abcdefghijklmnopqrstuvw',
-        'customer_code': 'CST1000',
-        'token_f4l4': '11119999',
-    }
-
-    transaction = gateway.BaseCardTransaction(
-        api_details=API_DETAILS, **details
-    )
-    status = transaction._determine_save_token_status(False)
-
-    assert status is False
 
 def test_determine_card_details_token():
     details = {
@@ -308,155 +238,3 @@ def test_determine_card_details_mag_encrypted_priority():
 
     assert len(transaction.cleaned) == 2
     assert 'mag_enc' in transaction.cleaned
-
-@patch(
-    'helcim.gateway.models.HelcimToken.objects.get_or_create',
-    mock_get_or_create_created
-)
-@patch.dict(
-    'helcim.gateway.SETTINGS',
-    {'redact_cc_name': False, 'redact_cc_expiry': False}
-)
-def test_save_token():
-    details = {
-        'token': 'abcdefghijklmnopqrstuvw',
-        'token_f4l4': '11119999',
-        'cc_name': 'Test Name',
-        'cc_expiry': '0120',
-        'customer_code': 'CST1000',
-    }
-
-    transaction = gateway.BaseCardTransaction(
-        api_details=API_DETAILS, django_user=1, **details
-    )
-    transaction.response = details
-    transaction.redact_data()
-    token_entry = transaction.save_token_to_vault()
-
-    # Checks that all proper fields ended up getting passed to model
-    assert token_entry.data['token'] == 'abcdefghijklmnopqrstuvw'
-    assert token_entry.data['token_f4l4'] == '11119999'
-    assert token_entry.data['cc_name'] == 'Test Name'
-    assert token_entry.data['cc_expiry'] == date(year=2020, month=1, day=31)
-    assert token_entry.data['customer_code'] == 'CST1000'
-    assert token_entry.data['django_user'] is None
-
-def test_save_token_missing_token():
-    details = {
-        'token_f4l4': '11119999',
-        'customer_code': 'CST1000',
-    }
-
-    transaction = gateway.BaseCardTransaction(
-        api_details=API_DETAILS, **details
-    )
-    transaction.response = details
-    transaction.django_user = 1
-    token_entry = transaction.save_token_to_vault()
-
-    # Checks that all proper fields ended up getting passed to model
-    assert token_entry is None
-
-def test_save_token_missing_token_f4l4():
-    details = {
-        'token': 'abcdefghijklmnopqrstuvw',
-        'customer_code': 'CST1000',
-    }
-
-    transaction = gateway.BaseCardTransaction(
-        api_details=API_DETAILS, django_user=1, **details
-    )
-    transaction.response = details
-    token_entry = transaction.save_token_to_vault()
-
-    # Checks that all proper fields ended up getting passed to model
-    assert token_entry is None
-
-def test_save_token_missing_customer_code():
-    details = {
-        'token': 'abcdefghijklmnopqrstuvw',
-        'token_f4l4': '11119999',
-    }
-
-    transaction = gateway.BaseCardTransaction(
-        api_details=API_DETAILS, **details
-    )
-    transaction.response = details
-    transaction.redact_data()
-
-    try:
-        transaction.save_token_to_vault()
-    except helcim_exceptions.ProcessingError as error:
-        assert str(error) == (
-            'Unable to save token - customer code not provided'
-        )
-    else:
-        assert False
-
-@patch(
-    'helcim.gateway.models.HelcimToken.objects.get_or_create',
-    mock_get_or_create_created
-)
-@patch.dict('helcim.gateway.SETTINGS', {'allow_anonymous': False})
-def test_save_token_with_django_user():
-    details = {
-        'token': 'abcdefghijklmnopqrstuvw',
-        'token_f4l4': '11119999',
-        'customer_code': 'CST1000',
-    }
-
-    transaction = gateway.BaseCardTransaction(
-        api_details=API_DETAILS, django_user=1, **details
-    )
-    transaction.response = details
-    token_entry = transaction.save_token_to_vault()
-
-    # Checks that all proper fields ended up getting passed to model
-    assert token_entry.data['token'] == 'abcdefghijklmnopqrstuvw'
-    assert token_entry.data['token_f4l4'] == '11119999'
-    assert token_entry.data['customer_code'] == 'CST1000'
-    assert token_entry.data['django_user'] == 1
-
-@patch.dict('helcim.gateway.SETTINGS', {'alllow_anonymous': False})
-def test_save_token_with_django_user_not_provided():
-    details = {
-        'token': 'abcdefghijklmnopqrstuvw',
-        'token_f4l4': '11119999',
-        'customer_code': 'CST1000',
-    }
-
-    transaction = gateway.BaseCardTransaction(
-        api_details=API_DETAILS, **details
-    )
-    transaction.response = details
-
-    try:
-        transaction.save_token_to_vault()
-    except helcim_exceptions.ProcessingError as error:
-        assert str(error) == 'Required Django user reference not provided.'
-    else:
-        assert False
-
-@patch(
-    'helcim.gateway.models.HelcimToken.objects.get_or_create',
-    mock_get_or_create_created
-)
-@patch.dict('helcim.gateway.SETTINGS', {'allow_anonymous': True})
-def test_save_token_with_customer_code():
-    details = {
-        'token': 'abcdefghijklmnopqrstuvw',
-        'token_f4l4': '11119999',
-        'customer_code': 'CST1000',
-    }
-
-    transaction = gateway.BaseCardTransaction(
-        api_details=API_DETAILS, **details
-    )
-    transaction.response = details
-    token_entry = transaction.save_token_to_vault()
-
-    # Checks that all proper fields ended up getting passed to model
-    assert token_entry.data['token'] == 'abcdefghijklmnopqrstuvw'
-    assert token_entry.data['token_f4l4'] == '11119999'
-    assert token_entry.data['customer_code'] == 'CST1000'
-    assert token_entry.data['django_user'] is None
